+++ bank/src/main/java/org/mixed/exam/bank/pojo/po/Paper.java
@@ -15,7 +15,7 @@ public class Paper
     private List<String> subjectIDs;//题目
     private String courseID;//试卷所属的学科
     private Boolean open=true;//试卷是否开放
-    private Integer difficulty=-1;//试卷难度
+    private Double difficulty=-1.0;//试卷难度
     private Date date=new Date();//试卷创建时间
     private Integer respondentCount=0;//答题人数
     private Integer correctCount=0;//答对人数
